+++ src/physical-layer-adapter/receive-adapter-state.service.js
@@ -3,7 +3,7 @@
     'use strict';
 
     AudioNetwork.Injector
-        .registerService('PhysicalLayer.ReceiveAdapterState', _ReceiveAdapterState);
+        .registerService('PhysicalLayerAdapter.ReceiveAdapterState', _ReceiveAdapterState);
 
     _ReceiveAdapterState.$inject = [];