+++ src/physical-layer-adapter/signal-power-collector/signal-power-collector-builder.service.js
@@ -3,10 +3,10 @@
     'use strict';
 
     AudioNetwork.Injector
-        .registerService('PhysicalLayer.SignalPowerCollectorBuilder', _SignalPowerCollectorBuilder);
+        .registerService('PhysicalLayerAdapter.SignalPowerCollectorBuilder', _SignalPowerCollectorBuilder);
 
     _SignalPowerCollectorBuilder.$inject = [
-        'PhysicalLayer.SignalPowerCollector'
+        'PhysicalLayerAdapter.SignalPowerCollector'
     ];
 
     function _SignalPowerCollectorBuilder(